pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "python-dotenv>=1.1.1",
    "pillow>=11.3.0",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "jwt>=1.4.0",
    "flask-limiter>=3.12",
]
